Switch to classic confinement
Classic confinement allows the snap to behave like a traditionally
packaged application with full access to the system, and enables the
use of traditional directories such as /etc and /var/log.

We will continue to store all of the snap's files in $SNAP* directories.
This enables the snap to cleanup after itself if it is removed. However,
traditional directory locations are symlinked to their corresponding
$SNAP* directories.

For example, glance configs are installed in $SNAP_COMMON/etc/glance
which has a symlink at /etc/glance.

The glance apps then use the traditional directories when running
commands and services.

Change-Id: Ic763f9c8f34ab3096e8592520db00cdf81b0d9ef